requirements.txt
PyYAML==3.12
pytest==3.2.1
pytest-catchlog==1.2.2
pyowm